feat: created testcase which verifies unconfirmed account can see own project
