+++ .github/workflows/run_testcases.yml
@@ -60,4 +60,4 @@ jobs:
           GCS_BUCKET: ${{ secrets.GCS_BUCKET }}
         run: |
           python manage.py migrate
-          pytest --cov=testing/unit
+          pytest --cov=test/unit
